06 - Resolved Conflicts
