+++ .gitignore
@@ -1,1 +1,1 @@
- 
+ /log/
